Deploying a demo amd64 debian-based OS via ostree
This is a prototype outline for building and deploying a debian based OS
with ostree.  For purposes of demonstration, the OS runs from a bootable
USB device on a UEFI capable amd64 machine.  Once a USB device is
created and booted, the OS can be updated using a simple shell script.

<https://ostreedev.github.io/ostree/>
<https://www.nin.wiki/Halo_numbers>